refactor(api): replace `useSend` with `use-send` for consistent naming
The `useSend` module has been renamed to `use-send` to align with naming conventions. Updated all imports and references accordingly. This change improves code readability and consistency across the codebase.
