+++ src/app/(social)/_popups/team-creation-popup.tsx
@@ -12,7 +12,7 @@ import {
 import { useUserInfo } from '@/lib/api/hooks/users';
 import { createTeamRequest } from '@/lib/api/models/team';
 import { ratelApi } from '@/lib/api/ratel_api';
-import { useApiCall } from '@/lib/api/useSend';
+import { useApiCall } from '@/lib/api/use-send';
 import { usePopup } from '@/lib/contexts/popup-service';
 import { logger } from '@/lib/logger';
 import { checkLowerAlphaNumeric } from '@/lib/valid-utils';
