+++ src/lib/api/use-send.ts
@@ -1,4 +1,5 @@
-// eslint-disable @typescript-eslint/no-explicit-any
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
 'use client';
 
 import { config } from '@/config';
